refactor relayhub
in order to deploy, relayhub must be broken down to separatte contracts.
also, must disable references to balance,
 and gasPrice (which is fixed)
(and to smaller extent - tx.origin)

build is based on hardhat, so did some modifications to run with
hardhat-truffle5
note: the hardhat-ovm breaks the hardhat-truffle integration, by
modifying the global "artifacts" object.
to fix that, a test should import "TruffleArtifact.js", which restore
that functionality (and also the top-level "contract" wrapper of
"describe")